Sun, May 19, 2024, 10:30 PM +06:00

+++ module-3/calc.cpp
@@ -2,7 +2,7 @@
 
 int main()
 {
-    int a = 20, b = 10, ans;
+    float a = 20, b = 10, ans;
 
     printf("give the value a : ");
     scanf(" %i", &a);
